+++ CHANGELOG.md
@@ -61,6 +61,7 @@ methods.
 * Removed local phone number validation when `giggsey/libphonenumber-for-php` was installed. Phone numbers are
   validated by the Firebase Service in any case, and even when a phone number was considered valid, in rare
   cases the Firebase API rejected them still.
+* Replaced `kreait/clock` with `beste/clock`, which implements the proposed [PSR-20 Clock Interface](https://github.com/php-fig/fig-standards/blob/master/proposed/clock.md).
 * The following classes and methods have been removed:
   * `Kreait\Firebase\Auth\ActionCodeSettings\RawActionCodeSettings`
   * `Kreait\Firebase\Project\ProjectId`